parsing-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pika==1.3.2
pymupdf==1.24.11
pdfplumber==0.11.4
openpyxl==3.1.5
python-docx==1.1.2
tiktoken==0.8.0
pydantic-settings==2.6.1